+++ app/src/main/java/com/example/c4q/capstone/LoginActivity.java
@@ -16,7 +16,7 @@ import android.view.MenuItem;
 import android.view.View;
 
 import com.example.c4q.capstone.database.model.publicuserdata.PublicUser;
-import com.example.c4q.capstone.network.NetworkCall;
+import com.example.c4q.capstone.network.barzz.NetworkCall;
 import com.example.c4q.capstone.userinterface.user.EditProfileActivity;
 import com.example.c4q.capstone.userinterface.user.SettingsActivity;
 import com.example.c4q.capstone.userinterface.user.UserProfileActivity;
